Leader/follower futures
Checking in a pair of futures about possible leader/follower
schemes to serve as conversation topics, and to give Steve
something to work towards.

These don't capture what I'd eventually like: to push coforalls
and on clauses into the leaders themselves and keep the invocation
of the leader simpler; but you've got to crawl before you walk...
(and a semantic question might be whether parallelism/on-clauses 
are only permitted within leader iterators, or whether they'd
be permitted within any iterator that can legally be inlined at 
all of its invocations).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14253 3a8e244f-b0f2-452b-bcba-4c88e055c3ca