Fixing a --gdb thinko/typo/bug
Steve was running into problems with the --gdb flag
again, which pointed out that after I'd gotten my
solution working yesterday, I continued to edit it
without testing, and introduced a bug into it.  In
particular, when I switched from using $_exitcode
to a custom convenience variable, I switched my
sentinel value of 777 to 1 in the sets of the
variable, but not in the tests.  This fixes that
bug.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14259 3a8e244f-b0f2-452b-bcba-4c88e055c3ca