Futures about A: [i in D] ... class members where D is also a member
It looks as though our forall initializers within a class
member don't work if the domain being iterated over is also
a class member.  This would be a useful feature for the
distribution code.  From appearances, it looks like we're
maybe not correctly determining which variables need to be
passed from the class scope into the function implementing
the forall expression/initializer.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14263 3a8e244f-b0f2-452b-bcba-4c88e055c3ca