Removed semantic future for test discussed at today's meeting and updated .good.  Fixed line number in error in related future and removed it.
About line numbers: We use a global variable currentLineno to set the
line number of new nodes in the IR.  If it is not set, the line number
we eventually report could be far off.  This changes sets this line
number in enough places so that this future works, but if anybody has
ideas for how to make this more systematic, that would be worthwhile.
One idea I've had is to always set and unset this global variable and
then flag an error if we end up building up IR nodes when it is unset.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14264 3a8e244f-b0f2-452b-bcba-4c88e055c3ca