Added order() and position() methods
In this change, I added an order() method to ranges and
domains, and a position() method to arithmetic domains.
Both methods take an index as their argument.  If the
index is within the range/domain, the order() method
will return its ordinal value using 0-based indexing:
0, 1, 2, 3, ...  For multidimensional domains, this
uses row-major order numbering.  If the index is not
a member, -1 cast to the range/domain's per-dimension
index type is returned.  Thus:

     (0..3).order(3) == 3
     (1..3).order(3) == 2
     (3..3).order(3) == 0
     (4..3).order(3) == -1
     (1..3 by 2).order(3) == 1
     ([1..3, 1..3]).order((2,1)) == 3
     etc.

position() is similar, but returns a tuple indicating
a per-dimension ordering of the index components.
Thus,

     ([1..3, 1..3]).position((2,1)) == (1,0)
     ([1..3, 1..3]).position((4,3)) == (-1,2)

I've found these functions useful in mapping from
a domain's indices to a 0-based numbering in
computing distributions in which a locale's ownership
is a function of its 0-based position in the array
of locales being passed in (where we can't, in
general assume that that array is 0-based).  I suspect
they (or something like them) will also be useful in
leader/follower iterators where we need to translate
things back to 0-based indices in order to have a
lingua franca between multiple distributions.

Added a few simple tests of the functions.

I'm not particularly crazy about the names of these
functions (though I don't think they're half bad either),
so feel free to propose improvements.  I couldn't find
an equivalent in Matlab (and didn't look too hard in 
Fortran, but did look a bit).

Possible future directions:

- have domain.order(i) take an optional type argument
  indicating the type of the return type should
  domain.numIndices overflow dim_type.

- permit an optional "origin" argument that would
  indicate where the counting should start (0 and 1
  being the most likely possibilities -- so could
  also use a bool if these were the only two cases
  we cared about).

   



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14266 3a8e244f-b0f2-452b-bcba-4c88e055c3ca