Next draft of 1D block distribution
This next draft of the 1D block distribution replaces the
associative domain over the locale type that I was previously
using to have each global class store an array of references
to the local classes with an array that uses the same domain
as the set of target locales passed into the distribution to
begin with.  As feared, this made the code less clean even
though it seems like the more logical way to store the array
of local classes.  We should discuss the pros and cons in our
next review of the distribution.

One question about this change is whether it clears up the
hacking on the locales array that David had to do in order
to get local { } working as he wanted within the domain
classes (I haven't seen his changes yet, so am not sure).

This affects the benchmarks themselves only in the sense
that they were written with knowledge of the local classes
in order to do the parallel initialization -- once leaders
and followers are available, we should be able to write
that initialization in a more distribution-neutral style.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14267 3a8e244f-b0f2-452b-bcba-4c88e055c3ca