In the previous revision I got ahead of myself and accidentally added:   modules/standard/ChapelRuntime.chpl
And modified the file:
  modules/standard/ChapelStandard.chpl

The ChapelRuntime module includes code written in Chapel that's intended
to be executed from the runtime module. In particular I'm using this module
to create a "table of tasks" structure that is much easier to manage
in Chapel than in C. This structure will be used by the runtime.

In order to keep the compiler in a working condition I'm adding a modification
to filesToAST.cpp which causes the compiler to generate code for the
ChapelRuntime standard module.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14270 3a8e244f-b0f2-452b-bcba-4c88e055c3ca