Started refactoring of armci/third-party Makefiles
Defined a bunch of directory locations related to
third-party stuff in make/Makefile.base and rewrote
the ARMCI Makefiles to avoid embedding so many
assumptions into the Makefiles.  The most specific
example of this was adding a utility script to the
armci directory that translates from our
CHPL_TARGET_PLATFORM setting to the requisite
ARMCI TARGET variable (e.g., linux -> LINUX,
sunos -> SOLARIS).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14275 3a8e244f-b0f2-452b-bcba-4c88e055c3ca