reenabled the single loop iterator optimization
With the new interface introduced in r14241, an iterator is typically
lowered to a class with the methods advance, hasMore, and
getValue so that the loop

  for i in foo() do ...

can be translated to

  var ic = foo();
  ic.advance();
  while (ic.hasMore()) {
    i = ic.getValue();
    ...
    ic.advance();
  }

The single loop iterator optimization creates additional methods
zip1, zip2, zip3, and zip4 so that the loop

  for i in foo() do ...

can be translated to

  var ic = foo();
  ic.zip1();
  while (ic.hasMore()) {
    ic.zip2();
    i = ic.getValue();
    ...
    ic.zip3();
  }
  ic.zip4();

If this optimization does not apply, zip1 and zip3 wrap advance and
zip2 and zip4 do nothing.  Otherwise, these methods get their code
from the four parts surrounding a loop and a yield as explained in
r10012.

Fixed a bug in the single loop iterator optimization in which live
variable analysis used to only apply to yield statements.  It now
applies to all spaces between the optimized iterator methods.  This
didn't seem to be an issue because user code always seems to read a
variable (use) when writing it (def) because an assignment is changed
into a move of a call to the assignment function which counts as a
use.  Since we now handle the iterator loops in the compiler after
resolution, we no longer necessary use a variable that is defined.
This bug would probably have surfaced later if, for example,
assignment functions are specialized and/or inlined in function
resolution.

Disabled the "c for loop" construct and primitive.  This is onerous to
deal with throughout compilation (and error prone) so I think it
should be reenabled just before code generation via some form of idiom
recognition.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14281 3a8e244f-b0f2-452b-bcba-4c88e055c3ca