Added hplExample1 study.
This is the version of HPL I went over with Brad and Steve in my weekly meeting
on July 8, 2008. This version is serial, blocked, and includes pivoting. It
does not not include the method used by HPL to verify results.  It includes a
number of testing functions to verify the implementation, these functions are
executed when the program is run.

This version uses the counted ranges feature we've been talking about (which
allows for range literals such as 3..#5). Included in this commit is a patch to
enable this feature.  To apply the patch move it to the $CHPL_HOME directory,
execute "patch -p0 < countedRange.patch", then recompile.

Some features of this version:

Some things we talked about in the meeting:
- I implemented the panel2D structure so that I'd have some way of
  passing a submatrix to a function where the submatrix was cognizant
  about what matrix it's embedded in. This is necessary for the panelSolve
  function which does a computation in a submatrix, but also needs to
  permute the rows of the full-matrix the submatrix is inside of.
  It's questionable whether this technique really improves readability of
  the code, so I'm going to try modifying this code to not have Panel2D
  and see how it goes.
- I'm unsure if the failure rate I'm getting for LU decomposition
  (about 20% is due to numerical stability issues or a bug in the
  algorithm).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14287 3a8e244f-b0f2-452b-bcba-4c88e055c3ca