Add a few tests of the chpldev_refToString function
classRef2String defines a second function like chpldev_refToString that takes
an object argument with blank intent. This function gives the address of the
object instead of the address of the pointer to the object.

domArrRef2String shows that domain and array variables can't have their
locations converted to strings, and Arr.domain can't either.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14290 3a8e244f-b0f2-452b-bcba-4c88e055c3ca