First working "launcher" using mpirun + Makefile refactorings
Established a new "launcher" concept which supports the idea of
wrapping a program like mpirun, aprun, load leveler, etc. into
a small front-end binary that does command line parsing and
then runs the "real" program, created as a second binary
(currently called, lamely and continuing in the ZPL tradition,
foo_real where foo is the program name -- suggestions for 
improvement appreciated).  Use two binaries for this purpose to
support the concept of running the launcher binary on the host
platform and the _real binary on the target platform (though
today both are compiled using the same settings... the target's).

The launcher command is determined by a script util/launcher.pl 
that makes a best decision as to what launcher to use based on
the communication layer and communication substrate (e.g.,
the conduit in GASNet or the choice of MPI, PVM, etc. in ARMCI).
Currently, it is set up to say "mpirun" for ARMCI running over
MPI and "none" in all other cases.  Eventually we'll want to
add support for more cases (like the XT's launch commands, apsim,
etc.).  As with other settings, the default can be overridden
using CHPL_LAUNCHER

I also added a util script commSubstrate.pl that returns the
communication substrate for a given platform.  Currently, for
gasnet it returns portals for cray-xt and udp otherwise (but
these values are not used anywhere, I believe), while for armci
it returns mpi, and for everything else it returns none.  It
can be overridden using the CHPL_COMM_SUBSTRATE variable.

Most of the launcher code is currently stored in src/main_launcher.c
which is a parallel to src/main.c.  It does the command-line
parsing and determines the number of locales, and then passes
this information along to a CHPL_LAUNCHER-specific routine that
converts it into a string that can be system()'d by main_launcher.c.
Each launcher's code is contained within a subdirectory of runtime 
named launch-$(CHPL_LAUNCHER)/chpl_cfg_launch.c.  
CHPL_LAUNCHER==none is special-cased in the Makefiles so that no
launcher files are created or compiled, and the only binary
created is the program itself.

The mpirun launcher is currently not as general as one would like,
hard-coding the path to mpirun into the source code (and assuming
platform == linux) rather than querying this information through
the utility scripts or some other mechanism.

The main challenge to getting the Makefiles working was to recompile
certain files that are used both for the launcher and real binaries
twice and have them end up in distinct places.  This motivated a
fair amount of the Makefile refactoring, to minimize the number of
places that had to change when various paths or strings are modified.
One remaining challenge that I did not undertake on this go-round is
to use different compilers/flags for the launcher as the real binary,
which is required if we want to support a host launcher and target
real binary.  This requires one more refactorization that was too
big to undertake in this check-in (and this check-in is useful on
its own terms because it permits easy experimentation with ARMCI).

Here are some of the other Makefile-related and incidental changes:

* Moved all directories containing make-generated files into a
  subdirectory named "gen" -- this gets away from having a
  subdirectory per source directory littering things up, and also
  makes "make clobber" work more as desired, blowing away anything
  created during the build process.

* Factored most definitions of directories into which objects are
  stored out of the runtime Makefiles and into new variables --
  RUNTIME_OBJDIR and LAUNCH_OBJDIR.

* Refactored Makefile.include and Makefiles in runtime subdirectories
  to minimize differences between them and factor variables into parent
  Makefiles

* Removed some duplicated -I paths that occurred when compiling
  generated code (and runtime files?)

* Removed some duplicated FORCE: rules

* Made top-level Makefile clobber once again remove all of lib/ and
  bin/ rather than just those directories corresponding to the current
  build.  Taking the definition of clobber to mean "return things to
  the raw subversion tree (or something quite close to that)", this
  seems most appropriate.  Maybe I'll change my mind again down the
  road and capture the rationale a little better...

* Added the name of the tmp/savec directory to the compiler-generated
  Makefile and cleaned up some of the code that named that and the
  temporary binary name.


Next steps:

- refactor top-level runtime Makefile so that different compiler
  settings/flags are used for the launcher and the _real libraries.

- generalize location of mpirun

- make GASNet build procedure use commSubstrate.pl script as part
  of its configuration steps.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14292 3a8e244f-b0f2-452b-bcba-4c88e055c3ca