Added support for array slicing via an out-of-bounds degenerate range
As proposed in Andy's meeting yesterday, added support for
slicing into an array using a partially-bounded range where
the given bound may be out of bounds.  Thus, given:

  var A: [1..10] real;

  A[11..] = ...

would be legal since it is equivalent to:

  A[11..10] = ...

and 11..10 describes a degenerate range/empty set of indices.
Therefore, the indices described by the range are not out
of bounds.

In an earlier checkin (r12490), Steve posed the question as to
whether A[11..12] should also result in an empty subarray.  My
argument is that it shouldn't because 11..12 is a non-empty
range and it contains indices that are not within A's domain.
(By contrast "11" is not part of 1..10, but it's not part of
11..10 either).

Updated the spec to discuss this case (and out-of-bounds
indexing in general, which seemed not to be discussed in
the obvious places).

Added a few new tests of this behavior.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14298 3a8e244f-b0f2-452b-bcba-4c88e055c3ca