Fixed a bug related to nested var iterators in which they would lose the reference to variables in an outer function.  Removed two of our older futures.
Fixed a bug in which a function could return an iterator.  Instead of
returning the iterator class, we really need to capture this into an
array first because the iterator may refer to local
variables/arguments which will no longer exist when the function
returns.  After fixing the first bug, this showed up in

  test/modules/figueroa/DuplicateConfigVars/RandomNumber.chpl

which is the only test that returns an iterator (by returning a loop
expression) as in

  def foo(n: int) {
    return for i in 1..n do i;
  }

If we return the iterator class, then it refers to the argument 'n'
which is on the stack.

Unfortunately, there are many generic internal functions in the
standard modules that we need to return iterator classes so there is a
significant conditional guarding the insertion of this returned array.
I'd like to think of a better way to do this; either by adding a
pragma to such internal functions or by disabling this insertion in
the standard modules.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14321 3a8e244f-b0f2-452b-bcba-4c88e055c3ca