Next draft of BradsBlock1DPar + unit test
This draft:

1) removes the solo case from the leader/follower iterator scheme,
   proposing instead that a these() iterator with no arguments
   should be the equivalent of solo -- why clutter it.

2) fixes the "failed to snapshot target locale set" issue that
   David pointed out last week and files a unit test to make
   sure that these semantics are preserved.

3) introduces a constructor to the distribution class to help
   make these semantics work -- this also removes the need for
   the initialize method.

4) introduces writeThis() debugging routines to the global and
   local distribution classes which are also used in the unit
   test above.

5) updates some comments that were out of date


This test also contains a commented out bug that I haven't yet
had the chance to capture into a future (and may not before
leaving town).  Search on "WANT TO DO" in the global distribution
class constructor.  For some reason, putting this loop into
the constructor rather than moving it out into a routine (which,
stylistically, I don't want to do) causes the compiler to try
and put a type onto the heap.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14327 3a8e244f-b0f2-452b-bcba-4c88e055c3ca