Add some destructor futures
The first shows that a destructor for class C can't be declared outside of the
body of C as: "def C.~C()".

The second shows that the tilde is not part of the name of the destructor, so
whitespace and comments are allowed between it and the name of the class.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14342 3a8e244f-b0f2-452b-bcba-4c88e055c3ca