added initial support for leader-follower iterators
Note: To enable parallel iterators, use --enable-parallel-iterators.
I had to use the flag because there are many tests that invoke
forall-loops on our standard arrays and domains yet we don't yet have
leader-follower iterators defined.  Without this flag, forall-loops
are handled as if they are for-loops.

Parallel iterators are specified by defining three iterators: a serial
iterator, a leader iterator, and a follower iterator.  The only
iterators supported until now were serial iterators.  Leader and
follower iterators must be defined identically in terms of their
arguments (same names, same types, same order---and this is checked)
except that the leader iterator must take an argument named 'leader'
and the follower iterator must take an argument named 'follower'.  The
leader argument is ignored; the follower argument is used to pass the
leader index to the follower iterator (or iterators---in zippered
iteration, only the followers are zippered; the first iterator is
executed in the leader context).

This is slightly different than all proposals so far; it turns out
that only the leader iterator needs to be distinguished in some way,
the follower iterator is marked by having the follower argument.  In
addition, the restriction that the arguments match exactly is
essential because we may not have a call site to the iterator by the
time we are examining the forall loop.  This is the case in promotion
or when an iterator call is the actual argument to a function that
takes a generic.

Specialized buildForallLoopStmt, used when parsing, to essentially
build up a coforall-loop over the leader and a zippered for-loop over
the followers.  Added _toLeader and _toFollower functions in the
standard modules as well as "to leader" and "to follower" primitives.
These functions are used to change iterator classes into leader
iterator classes and follower iterator classes respectively.  Calls to
these functions are inserted when we are building up the forall-loop.
The functions handle tuples so that the leader iterator becomes the
leader iterator for the first component of the tuple and the follower
iterator becomes a tuple of the follower iterators for each iterator.
The primitives are translated into functions that the compiler builds
up to change iterator classes into leader and follower iterator
classes.  Because we need to start with a class (not a call to an
iterator), I think it is essential that the formal arguments are
identical because it is only then that we can determine the mapping
from one class to another.  I think that without an identical
signature, the mapping could be different from one call to another
depending on whether named actuals were used in the call or not.

The leader and follower iterators are identified whenever an iterator
is called.  At this point, a bogus call to the leader and follower
iterators are constructed and resolution is attempted.  If resolution
succeeds, the iterator is marked and can be used in a forall-loop;
otherwise it cannot be.

Fixed a bug in which the coercion and ordered wrappers built for
iterators were changed into functions.  Since functions return arrays,
this could result in poor performance.  With the addition of
leader-follower iterators, this bug caused compilation issues because
a wrapped leader would no longer be an iterator (for example).

Added four test cases.  The (a) versions of these tests are written so
that the leader and follower variants iterate over a different space
so as to ensure that the serial iterator is not being called for both
loops.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14353 3a8e244f-b0f2-452b-bcba-4c88e055c3ca