Allow class destructors to be defined outside of its class definition
Also, removed the ability to define methods that overload operators.  This is
done by checking the name of the method during normalization.  Ideally, this
should be done while parsing, but I couldn't figure out how to do that.

Added a couple test cases related to the above, and removed David's future
that is now fixed with this check in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14357 3a8e244f-b0f2-452b-bcba-4c88e055c3ca