Improve the function to create homogeneous tuples
For tuples declared as "var t = paramN * type", a helper function in the "*"
function needed to be instantiated paramN times, increasing the size of the
tuple by one each time.

Instead, double the size of the tuple for each instantiation. This greatly
increases the size of tuples that can be created and improves compile time when
large tuples are involved.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14378 3a8e244f-b0f2-452b-bcba-4c88e055c3ca