Six variations on a test (motivated by Thom)
This is a set of 3 x 2 tests that try to pass
A+B in to a function expecting an array
argument via [blank, in, const] intent (the 3
axis).  This is crossed with passing in A+B
directly and passing a function "addem(A,B)"
that returns A+B.  The results of the current
compiler surprised me and raised questions
about semantics that I couldn't answer, so
filed 5 of 6 of these as the same semantic
future.  (the only one that worked as I
expected was arrArgIn-function.chpl).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14385 3a8e244f-b0f2-452b-bcba-4c88e055c3ca