added support for overriding iterator methods
Overriding iterators is supported by creating a subclass hierarchy
between iterator classes and casting between these types.  This change
extends iterator classes so that they contain an object 'super' field
allowing the dynamic type to be determined.  This dynamic type is used
to determine which iterator methods to call.  This dynamic dispatch is
done at the point when iterators are lowered.  It results in
significant complexity in the compiler and generated code but is
fairly orthogonal.

Refactored dynamic dispatch code to remove duplicated code for when
'this' is generic or not.

Robustified scalar replacement so that it still applies to iterator
classes despite having a dynamic ID tag.  We can ignore the setting of
this tag when doing the scalar replacement because if we are able to
scalar replace an instance of this class, then the tag will not be
necessary.

Note that overriding an iterator will seriously impact the performance
of the parent iterator since scalar replacement and iterator inlining
will not apply.

Removed four futures, fixing linefeeds in two of them by either adding
writeln statements to the program or removing newline characters in
the output.

Added five tests of overriding iterators.  The first three are basic
though two of them use iterators that result in structurally different
iterator classes.  The fourth zippers two dynamically dispatched
iterators.  The fifth checks the return type of overridden iterators
to make sure that they are equal or that there is a subtype
relationship.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14387 3a8e244f-b0f2-452b-bcba-4c88e055c3ca