Reworked the pthread version of the runtime to get rid of the completed field
Instead, when a task is about to begin executing, the pointer from the task pool
entry to the task list entry is nullified, as is the pointer from the task list
entry to the task pool entry.  This simplifies the code a bit, and as an added
bonus, it gets rid of a race condition in the previous version of this file.
Now, once a task begins executing its corresponding entry in both the task pool
as well as the task list can be freed at any time without any worries.

Also, chpl_begin no longer needs to return anything, which is an improvement
because what it used to return exposed a quirk of how the pthread version
was implemented.  The other threading versions were changed to match
chpl_begin's new interface.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14388 3a8e244f-b0f2-452b-bcba-4c88e055c3ca