Updating spec TODO, adding LONGTERM file
Updating the spec TODO with some items from Thom and others
I may have had sitting around.

Also, adding a LONGTERM file to the spec directory to capture
things that we retire as semantic futures, but indicated the
possibility of supporting in the long-term (once we've
conquered all the truly hard problems and just want to make
Chapel more fun to use).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14397 3a8e244f-b0f2-452b-bcba-4c88e055c3ca