Added support for counted ranges. Counted ranges are in the syntax:
<expr>..#<expr>

A counted range is a range from the value specified in the left hand side
expression plus the next n values where n is the value in the right hand side
expression.

For example: 3..#4 represents the range 3..6

For the time being the counting token (#) can only be applied to the right 
hand side of a range, and thus only the high bound may be counted.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14398 3a8e244f-b0f2-452b-bcba-4c88e055c3ca