more refactoring of function resolution
Added a file generics.cpp and moved code related to instantiation into
this file from functionResolution.cpp and symbol.cpp.

Added a file wrappers.cpp and moved code related to default-, order-,
coercion-, and promotion-wrappers into this file from
functionResolution.cpp and symbol.cpp.

Added a file caches.h and caches.cpp and moved code related to caches
for wrappers and instantiation into this file from symbol.cpp.

Added a file resolution.h to serve as a header file for
resolution-related functions and shared global variables.

This is a more functional organization and cuts in half the size of
symbol.cpp.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14400 3a8e244f-b0f2-452b-bcba-4c88e055c3ca