Record w/ array member test changes/additions
The original goal of the arrayInClass/*.chpl tests was to
see what happened to arrays in various guises when passed
to functions by blank intent, so here I'm backing out
Steve's check-in that converted one of the intents to inout
in order to preserve the .good file.  Given my current
understanding of records with array fields past by blank
intent (pass by copy-in + no local modifications permitted
rather than pass by reference + no local modifications
permitted), and the realization that one shouldn't be able
to modify such a record's array field when passing it by
blank intent (even though the current compiler permits it),
I commented out the lines that call baz() with a record
rather than make the test cause a problem later.  Updated
the .good files to reflect this change.

Also added a test to capture the "pass record with array
member by inout intent" case that Steve liked in his
rewrite of the tests (because this was one of the few
places that that was exercised).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14403 3a8e244f-b0f2-452b-bcba-4c88e055c3ca