Added hplExample3.
This snapshot of my HPL work uses a matrix-multiply written in mannar intended
for distributed parallelism. Specifically this matrix multiply comes from the
/studies/matrixMult/stonea/distributed/luLike2.chpl example, but it has been
modified to work with all sizes of matrix and block.

Added README file documenting the various versions of HPL in this directory.

Also removed countedRanges.patch since this patch was applied in revision
r14398.  Since this patch is now been applied hplExample1 can be included in
the nightly regression tests.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14407 3a8e244f-b0f2-452b-bcba-4c88e055c3ca