Add the beginnings of a block-cyclic distribution
This is an N-dimensional block-cyclic distribution with lots of current
limitations, the most obvious being the lack of parallel iterators and writer
methods.

Another limit is it forces domains to be one based, and forces the array of
locales it's distributed over to be zero based.

However, it is working to the point that a commit seems reasonable.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14409 3a8e244f-b0f2-452b-bcba-4c88e055c3ca