Filing some new Thom futures
Thom was trying to capture an unknown number of ints into a class.
Took a stab at a few ways of doing this, two of which worked, one
of which led to a semantic future for me.  See the .future file
for details (related to what the "safe" initialization of an
array member should mean in the context of a user constructor).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14412 3a8e244f-b0f2-452b-bcba-4c88e055c3ca