Implemented variable-width boolean types; mapped bool to _Bool
On my plane flight home the other week I didn't have the FAQ I
wanted to work on and did have a week-stale snapshot of the
Chapel compiler, so decided to work on something constructive
but unrelated to the main things I was working on, and chose
booleans, since they have had a performance impact on tests
like Franz' FFT tests.  This was also a good exercise for
forcing me to get a little bit more familiar with the function
resolution code.

Since we discovered this performance gap, I've been taking the
stance that we should support the "bool" type to have an
implementation-defined width, while also supporting "bool(n)"
types.  Unlike int types, the bool(n) type does not affect
the values that can be stored in the type, only the amount of
memory used to store the bool.  Thus, an array of bool(8)s
would use far less memory than an array of bool(64)s.  This
would allow a programmer to make time-space tradeoffs if need
be.

Implemented this functionality for bools of size 8, 16, 32, and
64, and made our default "bool" type map to C99's _Bool type,
under the assumption that that was going to be portable and
efficient (i.e., hopefully the C implementors have made a wise
time-space tradeoff and also optimized the _Bool implementation
when possible).

Since all bools store the same values, I made all bool types
implicitly coerce to one another.  This is attractive because
it prevents us from needing to rewrite the standard module
routines that accept a bool -- they can still accept a bool
and be passed any of the available widths.

For awhile, I added a generic "boolean" type like our "integral"
type, but then I realized that because of the implicit coercions,
this isn't particularly useful (that I could think of) and was
likely to only add overhead in the form of code replicants.

I added a bunch of tests to make sure that bools of different
widths could interassign, interoperate, and be applied to all
the standard operators correctly.  I also added some tests to
see what would happen when functions are overloaded on bools,
and found that it worked as you would expect -- if there is
only one match, any bool will go to it; if there is an exact
match, the bool will go to it; if there are multiple imprecise
matches, it is ambiguous.

Also found that ~ doesn't work on param bools (unrelated to my
changes), so filed a pair of futures against this for future
work.

Updated the spec.

Implementation details:

* Changes to ChapelBase:
  - added an _isBooleanType funtion to ChapelBase.chpl and refined the
    definitions of other _is*Type functions to accept bools of various
    widths

  - added _cast() functions for all the bool widths (it seemed that
    this was one case where the "bool" version didn't work for all
    bool widths -- do casts not get implicit conversions?  (seems
    plausible)

  - changed instances of "== bool" to _isBooleanType()

* Added an array of PrimitiveType* called dtBools that's similar to
  dtInt, dtUInt, except that it has an additional element called
  BOOL_SIZE_SYS which corresponds to the system-chosen size of the
  boolean.  I kept the dtBool PrimitiveType* around for convenience
  and made it alias the same PrimitiveType* as the BOOL_SIZE_SYS
  entry, since it gets old typing "dtBools[BOOL_SIZE_SYS]" every time
  you need a boolean type.

* Changed comparisons against "== dtBool" into "is_bool_type()"
  function calls and implemented this function.  Comparisons against
  dtBool should rarely be used (in fact, there are currently none).

* implemented the implicit conversion rules in functionResolution.cpp

* implemented code which seems to treat "bool" as a generic argument
  type, just as "int" seems to be later (in the function
  (clone_parameterized_primitive_methods), but I'm not sure that any
  of this matches our current notion of the language.  Originally, I
  think we considered int to be similar to what "int(?)" is now, but
  I'm not sure that this is still the case.  Was surprised to stumble
  across this code.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14417 3a8e244f-b0f2-452b-bcba-4c88e055c3ca