Filing bug to capture problem that Block1DPar works around
After thinking that the thing I was working around had miraculously
fixed itself, found that I'd forgotten that it was a --no-local
thing only.  Put the workaround back in and filed a bug against
the problem which seems to be related to constructing a generic
class within an on clause within an iterator within a constructor.
(well, I'm less sure about the iterator part, but the rest seems
to be true).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14418 3a8e244f-b0f2-452b-bcba-4c88e055c3ca