Fix some issues with delete and destructors
Using delete on an object with its nearest matching destructor being the one
declared for the object class was resulting in a runtime infinite recursion.
This was because the ~object destructor had a call to the chpl_free primitive
which gets translated into a call to the destructor.

Previously, "delete c;" was translated in the parser to: "chpl_free(c)" and
later all calls to "chpl_free(c)" are translated to "chpl_destroy(c)" with
"chpl_destroy" being the name of every destructor. All calls to "chpl_free"
go away in this transformation so no uses of delete were freeing any memory.
I changed this to call "chpl_destroy(c)" then "chpl_free(c)".

Identifiers, including class names, are allowed to use the '$' character. When
checking that a destructor is named after the class it is defined in, compare
the destructor's declared name with the name of the type of its 'this'
argument. Once the name is known to match, change it to "~chpl_destroy". I
added the '~' so it won't clash with user identifiers.

With --no-local, pass the .addr field of wide classes to chpl_free. Previously
this wasn't an issue since chpl_free wasn't called.

Added some now passing delete tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14423 3a8e244f-b0f2-452b-bcba-4c88e055c3ca