Removed open interval syntax
Rejoice!

Removed the [expr[, expr]*) syntax from the compiler
now that lo..#hi is implemented, removed its description
from the spec, and rewrote the tests that used it (and
removed tests whose only purpose was to test it).

Here is a brief summary of how I rewrote things:

[0..hi)  => 0..#hi            (if a range was desired)
         => [0..#hi]          (if a domain was desired)

[lo..hi) => lo..hi-1          (if a range was desired)
            [lo..hi-1]        (if a domain was desired)

[D) => [D.low..D.high-1]      (I think we may want chomp(D) or D.chomp() 
                               a la Perl for cases like this one)

off + [0..hi) => off..#hi     NICE!

[lo..lo+num)  => lo..#num     NICE!

[0..radix)*str + lo => (0..#radix)*str + lo

Note that this last case (from the FFT codes) wants the ability
to do what we were discussing in our adoption of lo..#hi -- the
ability to say "give me 4 elements counting by 2s."  At the same
time, other expressions in FFT wanted the opposite interpretation
"count 4 elements and then stride the result by 2."  This suggests
to me that the "make by both a ternary and a binary operator"
approach isn't ideal since it would require 1..#num by 2 for the
former and (1..#num) by 2 for the latter -- in particular, the
parenthesization in the second case seems like the natural
parenthesization for the original case, which is ugly.

Over lunch, Andy suggested the reverse:  make 1..#num by 2 be
what it currently is and support 1..(#num by 2) as the way of
saying "count of num elements, counting by 2".  This seems
really natural and intuitive to me, so I suggested that he
look into implementing it if he has time.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca